VueRouter: Add StartPageSearchPanel component to make abstraction from base component and control routing parameters, update route entrypoint for basic route with search panel
